[SPARK-20417] Move subquery error handling to checkAnalysis from Analyzer.
